pyproject.toml
[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false    # TODO: eliminate errors and enable at CI
mypy_preset = "entry"  # TODO: "strict"
line_length = 88
py_version = "py38"
isort_known_first_party = ["voicevox_engine"]
isort_known_third_party = ["numpy"]
[[tool.pysen.lint.mypy_targets]]
  paths = [".", "voicevox_engine/"]

[tool.black] # automatically generated by pysen
line-length = 88
target-version = ["py38"]


[tool.isort] # automatically generated by pysen
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
known_first_party = ["voicevox_engine"]
known_third_party = ["numpy"]
line_length = 88
multi_line_output = 3
use_parentheses = true

[tool.poetry]
name = "voicevox_engine"
version = "0.0.0"
description = ""
authors = ["Hiroshiba <hihokaruta@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
numpy = "^1.20.0"
fastapi = "^0.70.0"
python-multipart = "^0.0.5"
uvicorn = "^0.15.0"
aiofiles = "^0.7.0"
soundfile = "^0.10.3.post1"
scipy = "^1.7.1"
pyyaml = "^6.0"
pyworld = "^0.3.0"
appdirs = "^1.4.4"
requests = "^2.28.1"
jinja2 = "^3.1.2"
pyopenjtalk = {git = "https://github.com/VOICEVOX/pyopenjtalk", rev = "f4ade29ef9a4f43d8605103cb5bacc29e0b2ccae"}
espnet = {git = "https://github.com/voicevox-bridge/espnet", rev = "d7633eb0eec9b00646847abdee3d34e5a62d2dea"}

[tool.poetry.group.dev.dependencies]
cython = "^0.29.24"
pre-commit = "^2.16.0"
atomicwrites = "^1.4.0"
colorama = "^0.4.4"
poetry = "^1.3.1"
nuitka = "^1.3.7"

[tool.poetry.group.test.dependencies]
pysen = {version = "~0", extras = ["lint"]}
pytest = "^6.2.5"
coveralls = "^3.2.0"

[tool.poetry.group.license.dependencies]
pip-licenses = "^3.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
